Add cli option to do migration
The migration is quite specific to the DTR project, and a few
things are hard coded. The reasoning behind this is that the activity of
migrating is quite limited so it doesn't make that much sense to invest
large amounts of time in writing a more generic migration script.

The script takes a directory as an argument. It traverses the directory
and for each file which represents a test run it creates the
corresponding test run in the Mirantis test rail account, with the
corresponding results.

Signed-off-by: Jose Bigio <jose.bigio@docker.com>